+++ src/world7.cpp
@@ -141,7 +141,7 @@ namespace World7 {
         BossState& currentBoss  = Bosses::ActiveBoss(game);
 
         //NOTE this is incorrect (Rectangle takes for paramters
-        Rectangle BossBox = { (int)boss_x, (int)boss_y, boss_scale };
+        Rectangle BossBox = { (float)boss_x, (float)boss_y, (float)boss_scale };
         //use the built in get boss hitbox , will return rectangle for you
         if (CheckCollisionCircleRec(ballPosition, ballRadius, BossBox)) {
             std::cout << "HIT!\n";
@@ -149,7 +149,7 @@ namespace World7 {
             ballPosition = { 10, 10 };
         }
         if (CheckCollisionCircleLine(ballPosition, ballRadius, collsionLinePoint1, collsionLinePoint2)){
-            Vector2 bossPos = { boss_x, boss_y };
+            Vector2 bossPos = { (float)boss_x, (float)boss_y };
             Vector2 directionToBoss = Vector2Subtract(bossPos, ballPosition);
             ballSpeed = Vector2Normalize(directionToBoss);
             ballSpeed = Vector2Scale(ballSpeed, initialBallSpeed);
